requirements.txt
pandas
pymongo
python-dotenv
fastapi
uvicorn
unidecode